pyproject.toml
[project]
name = "stylish-onnx"
version = "0.1.0"
description = "TTS with Stylish (StyleTTS2) and onnxruntime"
readme = "README.md"
authors = [
    { name = "thewh1teagle", email = "61390950+thewh1teagle@users.noreply.github.com" }
]
requires-python = ">=3.11"
dependencies = [
    "espeakng-loader>=0.2.4",
    "numpy>=2.3.0",
    "onnxruntime>=1.22.0",
    "phonemizer-fork>=3.3.2",
    "soundfile>=0.13.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
